+++ CHANGELOG.md
@@ -11,3 +11,13 @@ cái này để search, gồm các tham số
 POST http://localhost/api/newsletter/index
 
 cái url này để POST dữ liệu để tạo index (xem tools/create_index.py)
+
+# 20/11/2021: @tuana9a đã update lại tham số truyền vào
+
+GET http://localhost/api/newsletter/search?field=content&top=1&q=code
+
+cái này để search, gồm các tham số
+
+- field: trường muốn search VD: content, url, tittle
+- top: giới hạn số bản ghi trả về (top K trong lý thuyết)
+- q: là câu truy vấn